Cleaned up folder-structure
This Commit cleans up the folder structure to have listeners and events
in dedicated folders. Also the namespace has changed to ease difference
between event (conference) and event (pubsub-notice).
