+++ src/pubsub/Event/AutoApprovedEventCreated.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Joindin\Events;
+namespace Joindin\Pubsub\Event;
 
 class AutoApprovedEventCreated extends AbstractEvent
 {
