+++ src/pubsub/EventInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Joindin\Pubsub;
+
+interface EventInterface
+{
+    /**
+     * @return string
+     */
+    public static function getEventName();
+
+    public function getName();
+}
